requirements.txt
dash
requests
pandas
python-dotenv
plotly
gunicorn